+++ host/m5resolver/__init__.py
@@ -1,7 +1,8 @@
 from .agent import AgenticController, DeviceState
-from .controller import IntentController, TelemetryFrame
+from .controller import IntentController, TelemetryFrame, TimeTravelFrame
 from .fluxwire import ContinuousWire, FluxGraph, MeshBus
 from .mesh import FluxwireGossipMesh
+from .replay_engine import HostReplayEngine, ReplayResult, TIME_TRAVEL_PREFIX
 from .registry import DriverRegistry
 from .registry_ops import RegistryStore
 from .simulation import HardwareSimulator
@@ -13,10 +14,14 @@ __all__ = [
     "DeviceState",
     "IntentController",
     "TelemetryFrame",
+    "TimeTravelFrame",
     "ContinuousWire",
     "FluxGraph",
     "MeshBus",
     "FluxwireGossipMesh",
+    "HostReplayEngine",
+    "ReplayResult",
+    "TIME_TRAVEL_PREFIX",
     "DriverRegistry",
     "RegistryStore",
     "HardwareSimulator",